Removed all user profile related operations out of the suer service into a stand-alone user profile service. Updated all the affected user controller and service unit tests. Fixed a bug in the user service where the ITransactable methods did not reference all the repositories used within the user service.
